% File src/library/grid/man/grid-internal.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2013 R Core Team
% Distributed under GPL 2 or later

\name{grid-internal}
%% Part of the sensible export list but (currently?) documented as
%% internal (or waiting for documentation to be written):
%% <LIST>
\alias{makeinter_cpp}
\alias{makethreeinter_cpp}
\alias{updatebeta_cpp}
\alias{updateREs}
\alias{make.inter}
\alias{make.inter.treat}
\alias{make.justthreeway}
\alias{make.threewayinter}

\title{Internal Sparsereg Functions}
\description{
  Internal sparsereg functions
}
\details{
  These are not to be called by the user.
}
\keyword{ internal }
