\name{sparsereg-package}
\alias{sparsereg-package}
\docType{package}
\title{
Sparse regression for experimental and observational data.
}
\description{
Sparse modeling provides a mean selecting a small number of non-zero effects from a large possible number of candidate effects.  This package includes a suite of methods for sparse modeling: estimation via EM or MCMC, approximate confidence intervals with nominal coverage, and diagnostic and summary plots.  Beyond regression analyses, applications include subgroup analysis, particularly for conjoint experiments, and panel data. Future versions will include extensions to limited dependent variables, models with truncated outcomes, and propensity score and instrumental variable analysis.

}
\details{
\tabular{ll}{
Package: \tab sparsereg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-20\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Marc Ratkovic and Dustin Tingley
Maintainer: Marc Ratkovic (ratkovic@princeton.edu)
}
\references{
Ratkovic, Marc and Tingley, Dustin.  2015.  "Sparse Estimation with Uncertainty: Subgroup Analysis in Large Dimensional Design."  Working paper.
}
\keyword{ package }
\seealso{
\code{FindIt}, \code{glmnet}
}
